Move screens into pages
